Add files via upload
Old Today , 12:10   [L4D/L4D2] Teleport survivors (v2.6, 2025-11-27)
Edit/Delete Message Reply With Quote Multi-Quote This Message Quick reply to this message #1
This plugin teleports lagging players to elevators, shelters, and rescue vehicles. Other incapacitated players die.

Thanks to alasfourom (https://forums.alliedmods.net/showthread.php?p=2785701) for the elevator idea, which got me thinking about finding a universal automatic way to detect elevator buttons that would work in both official maps and add-ons. It looks like I've found a way to do jthat, and so I decided to create a plugin that will integrate teleportation to elevators, safe rooms, rescue vehicles, and perhaps some other similar scenarios.

Teleportation of lagging players is activated in the following cases:
- when pressing the elevator button
- when closing the shelter door
- when entering the rescue vehicle

Teleportation is activated when the minimum number of players, as defined by the corresponding variable, is reached. A countdown timer is started, after which the players are transported and the elevator is started or the safe room door is closed. If the number of players falls below this variable or all players are present, teleportation is cancelled.